+++ lib/fortnite_api_io.dart
@@ -3,3 +3,5 @@ export "src/fortnite_api_io/fortnite_api_io.dart";
 export "src/fortnite_api_io/structures/challenges.dart";
 export "src/fortnite_api_io/structures/items.dart";
 export "src/fortnite_api_io/structures/item_sets.dart";
+export "src/fortnite_api_io/structures/rarities.dart";
+export "src/fortnite_api_io/structures/shop.dart";